requirements.txt
flask
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http